+++ example/lib/demo/screens/color_picker/color_picker_card.dart
@@ -48,6 +48,8 @@ class ColorPickerCard extends ConsumerWidget {
           secondaryMenu: watch(secondaryMenuPod).state,
           secondaryOnDesktopLongOnDevice:
               watch(secondaryDesktopOtherLongPod).state,
+          secondaryOnDesktopLongOnDeviceAndWeb:
+              watch(secondaryDesktopWebLongPod).state,
           editFieldCopyButton: watch(editFieldCopyButtonPod).state,
           parseShortHexCode: watch(parseShortHexCodePod).state,
           editUsesParsedPaste: watch(editUsesParsedPastePod).state,